backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
python-dateutil
google-generativeai
python-dotenv
feedparser
requests
